yabc cmd and docs
