+++ src/yabc/app/__main__.py
@@ -1,0 +1,4 @@
+import yabc
+
+app = yabc.app.create_app()
+app.run()
