+++ src/yabc/formats/__init__.py
@@ -1,5 +1,3 @@
-from . import adhoc
-from . import coinbase
-from . import gemini
-
-format_classes = [adhoc.AdhocParser, coinbase.CoinbaseParser, gemini.GeminiParser]
+FORMAT_CLASSES = []
+__all__ = ["adhoc", "gemini", "coinbase"]
+from . import *
