+++ src/yabc/formats/coinbase.py
@@ -4,8 +4,12 @@ Code for importing data from a report generated by the exchange Coinbase.
 YABC is not affiliated with the company Coinbase or any of its products.
 """
 import csv
+import decimal
+
+import dateutil
 
 from yabc import transaction
+from yabc.formats import FORMAT_CLASSES
 
 
 def from_coinbase(f):
@@ -46,7 +50,7 @@ def txs_from_coinbase(f):
     :return:
     """
     dicts = from_coinbase(f)
-    return [transaction.Transaction.FromCoinbaseJSON(i) for i in dicts]
+    return [FromCoinbaseJSON(i) for i in dicts]
 
 
 class CoinbaseParser:
@@ -64,3 +68,35 @@ class CoinbaseParser:
         if not self.txs:
             raise StopIteration
         return self.txs.pop(0)
+
+def FromCoinbaseJSON(json):
+    """
+    Arguments:
+        json (dict): a coinbase-style dictionary with the following fields:
+            - 'Transfer Total': the total USD price, including fees.
+            - 'Transfer Fee': the USD fee charged by the exchange.
+            - 'Amount': the amount of bitcoin sold. (It's negative for sales.)
+            - 'Currency': which cryptocurrency was involved.
+            - 'Timestamp': 'hour:min:sec.millisecs' formatted timestamp.
+    Returns: Transaction instance with important fields populated
+    """
+    operation = transaction.Transaction.Operation.BUY
+    proceeds = json["Transfer Total"]
+    fee = json["Transfer Fee"]
+    asset_name = json["Currency"]
+    quantity = decimal.Decimal(json["Amount"])
+    if quantity < 0:
+        operation = transaction.Transaction.Operation.SELL
+        quantity = abs(quantity)
+    timestamp_str = json["Timestamp"]
+    return transaction.Transaction(
+        asset_name=asset_name,
+        operation=operation,
+        quantity=quantity,
+        date=dateutil.parser.parse(timestamp_str),
+        fees=fee,
+        source="coinbase",
+        usd_subtotal=proceeds,
+    )
+
+FORMAT_CLASSES.append(CoinbaseParser)
