+++ tests/csv_test.py
@@ -5,9 +5,9 @@ Test that we can parse data from various CSV sources.
 import glob
 import unittest
 
-from yabc import csv_to_json
 from yabc.formats import adhoc
 from yabc.formats import coinbase
+from yabc.formats import gemini
 
 
 class CsvTest(unittest.TestCase):
@@ -22,7 +22,7 @@ class CsvTest(unittest.TestCase):
 
     def test_load_gemini_csv(self):
         """ Test wholesale load of gemini data from CSV"""
-        csv_to_json.gemini_to_dict("testdata/synthetic_gemini_csv.csv")
+        gemini.gemini_to_dict("testdata/synthetic_gemini_csv.csv")
 
     def test_load_adhoc_csv(self):
         """ Test wholesale load of gemini data from CSV"""
