+++ tests/transaction_test.py
@@ -10,6 +10,7 @@ from yabc import Base
 from yabc import basis
 from yabc import transaction
 from yabc import user  # noqa
+from yabc.formats import gemini
 from yabc.transaction import Transaction
 from yabc.transaction import make_transaction
 
@@ -136,7 +137,7 @@ class TransactionTest(unittest.TestCase):
             # "Time": "06:27:56.373",
         }
 
-        trans = transaction.Transaction.FromGeminiJSON(gemini_json_buy)
+        trans = gemini.FromGeminiJSON(gemini_json_buy)
 
         self.assertEqual(trans.operation, BUY)
         self.assertEqual(trans.quantity, 2)
@@ -159,7 +160,7 @@ class TransactionTest(unittest.TestCase):
             # "Time": "06:27:56.373",
         }
 
-        trans = transaction.Transaction.FromGeminiJSON(gemini_json_sell)
+        trans = gemini.FromGeminiJSON(gemini_json_sell)
 
         self.assertEqual(trans.operation, SELL)
         self.assertEqual(trans.quantity, 2)
@@ -180,7 +181,7 @@ class TransactionTest(unittest.TestCase):
             "Date": "2015-2-5",
         }
 
-        trans = transaction.Transaction.FromGeminiJSON(sell_json_gemini)
+        trans = gemini.FromGeminiJSON(sell_json_gemini)
         engine = sqlalchemy.create_engine("sqlite:///:memory:", echo=True)
         Session = sessionmaker(bind=engine)
         session = Session()
